update distance utils

+++ domain/src/Scenario.jl
@@ -16,7 +16,7 @@ struct Scenario
     bufferTime::Int # Latest call time for a request in minutes
     maximumDriveTimePercent::Int # Percent of direct drive time to find maximum ride time
     minimumMaximumDriveTime::Int # Minimum duration in minutes of maximum drive time 
-    distance::Array{Int, 2}
+    distance::Array{Float64, 2}
     time::Array{Int, 2}
     nDepots::Int
     depots::Dict
